third_party/blink/web_tests/external/wpt/shadow-dom/selection-direction.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] direction returns "none" when there is no selection
  assert_equals: expected (string) "none" but got (undefined) undefined
[FAIL] direction returns "forward" when there is a forward-direction selection in the document tree
  assert_equals: expected (string) "forward" but got (undefined) undefined
[FAIL] direction returns "backward" when there is a backward-direction selection in the document tree
  assert_equals: expected (string) "backward" but got (undefined) undefined
[FAIL] direction returns "forward" when there is a forward selection in the shadow tree
  assert_equals: expected (string) "forward" but got (undefined) undefined
[FAIL] direction returns "backward" when there is a backward selection in the shadow tree
  assert_equals: expected (string) "backward" but got (undefined) undefined
[FAIL] direction returns "forward" when there is a forward selection that crosses shadow boundaries
  assert_equals: expected (string) "forward" but got (undefined) undefined
[FAIL] direction returns "backward" when there is a forward selection that crosses shadow boundaries
  assert_equals: expected (string) "backward" but got (undefined) undefined
Harness: the test ran to completion.
